devtools/client/shared/components/reps/test/mochitest/test_reps_grip-array.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test GripArray rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - GripArray</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  const {
    REPS,
    MODE,
    getRep,
    getSelectableInInspectorGrips,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { GripArray } = REPS;

  let componentUnderTest = GripArray;
  const maxLength = {
    short: 3,
    long: 10
  };

  try {
    yield testBasic();

    // Test property iterator
    yield testMaxProps();
    yield testMoreThanShortMaxProps();
    yield testMoreThanLongMaxProps();
    yield testRecursiveArray();
    yield testPreviewLimit();
    yield testEmptySlots();
    yield testNamedNodeMap();
    yield testNodeList();
    yield testDocumentFragment();

    yield testOnDomNodeMouseOver();
    yield testOnDomNodeMouseOut();
    yield testOnDomNodeInspectIconClick();

    yield testObjectLink();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testBasic() {
    // Test array: `[]`
    const testName = "testBasic";

    // Test that correct rep is chosen
    const gripStub = getGripStub("testBasic");
    is(getRep(gripStub), GripArray.rep, "Rep correctly selects GripArray Rep");

    // Test rendering
    const defaultOutput = `Array []`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testMaxProps() {
    // Test array: `[1, "foo", {}]`;
    const testName = "testMaxProps";

    const defaultOutput = `Array [ 1, "foo", Object ]`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[3]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      },
      {
        // Check the custom title with nested objects to make sure nested objects are not
        // displayed with their parent's title.
        mode: MODE.LONG,
        title: "CustomTitle",
        expectedOutput: `CustomTitle [ 1, "foo", Object ]`,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testMoreThanShortMaxProps() {
    // Test array = `["test string"…] //4 items`
    const testName = "testMoreThanShortMaxProps";

    const defaultOutput = `Array [ ${Array(maxLength.short).fill("\"test string\"").join(", ")}, 1 more… ]`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[${maxLength.short + 1}]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: `Array [ ${Array(maxLength.short + 1).fill("\"test string\"").join(", ")} ]`,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testMoreThanLongMaxProps() {
    // Test array = `["test string"…] //301 items`
    const testName = "testMoreThanLongMaxProps";

    const defaultShortOutput = `Array [ ${Array(maxLength.short).fill("\"test string\"").join(", ")}, ${maxLength.long + 1 - maxLength.short} more… ]`;
    const defaultLongOutput = `Array [ ${Array(maxLength.long).fill("\"test string\"").join(", ")}, 1 more… ]`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultShortOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[${maxLength.long + 1}]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultShortOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultLongOutput
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testRecursiveArray() {
    // Test array = `let a = []; a = [a]`
    const testName = "testRecursiveArray";

    const defaultOutput = `Array [ [1] ]`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[1]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testPreviewLimit() {
    const testName = "testPreviewLimit";

    const shortOutput = `Array [ 0, 1, 2, 8 more… ]`;
    const defaultOutput = `Array [ 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 1 more… ]`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: shortOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[11]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: shortOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testEmptySlots() {
    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ <5 empty slots> ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[5]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ <5 empty slots> ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ <5 empty slots> ]",
      }],
      "Array with empty slots only",
      componentUnderTest,
      getGripStub("Array(5)")
    );

    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ <1 empty slot>, 1, 2, 1 more… ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[4]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ <1 empty slot>, 1, 2, 1 more… ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ <1 empty slot>, 1, 2, 3 ]",
      }],
      "Array with one empty slot at the beginning",
      componentUnderTest,
      getGripStub("[,1,2,3]")
    );

    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ <3 empty slots>, 3, 4, 1 more… ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[6]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ <3 empty slots>, 3, 4, 1 more… ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ <3 empty slots>, 3, 4, 5 ]",
      }],
      "Array with multiple consecutive empty slots at the beginning",
      componentUnderTest,
      getGripStub("[,,,3,4,5]")
    );

    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ 0, 1, <1 empty slot>, 3 more… ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[6]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ 0, 1, <1 empty slot>, 3 more… ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ 0, 1, <1 empty slot>, 3, 4, 5 ]",
      }],
      "Array with one empty slot in the middle",
      componentUnderTest,
      getGripStub("[0,1,,3,4,5]")
    );

    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ 0, 1, <3 empty slots>, 1 more… ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[6]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ 0, 1, <3 empty slots>, 1 more… ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ 0, 1, <3 empty slots>, 5 ]",
      }],
      "Array with multiple successive empty slots in the middle",
      componentUnderTest,
      getGripStub("[0,1,,,,5]")
    );

    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ 0, <1 empty slot>, 2, 3 more… ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[6]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ 0, <1 empty slot>, 2, 3 more… ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ 0, <1 empty slot>, 2, <1 empty slot>, 4, 5 ]",
      }],
      "Array with multiple non successive single empty slots",
      componentUnderTest,
      getGripStub("[0,,2,,4,5]")
    );

    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ 0, <2 empty slots>, 3, 5 more… ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[9]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ 0, <2 empty slots>, 3, 5 more… ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ 0, <2 empty slots>, 3, <3 empty slots>, 7, 8 ]",
      }],
      "Array with multiple multi-slot holes",
      componentUnderTest,
      getGripStub("[0,,,3,,,,7,8]")
    );

    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ 0, 1, 2, 3 more… ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[6]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ 0, 1, 2, 3 more… ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ 0, 1, 2, 3, 4, <1 empty slot> ]",
      }],
      "Array with a single slot hole at the end",
      componentUnderTest,
      getGripStub("[0,1,2,3,4,,]")
    );

    testRepRenderModes(
      [{
        mode: undefined,
        expectedOutput: "Array [ 0, 1, 2, 3 more… ]",
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[6]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: "Array [ 0, 1, 2, 3 more… ]",
      },
      {
        mode: MODE.LONG,
        expectedOutput: "Array [ 0, 1, 2, <3 empty slots> ]",
      }],
      "Array with multiple consecutive empty slots at the end",
      componentUnderTest,
      getGripStub("[0,1,2,,,,]")
    );
  }

  function testNamedNodeMap() {
    const testName = "testNamedNodeMap";

    const defaultOutput = `NamedNodeMap [ class="myclass", cellpadding="7", border="3" ]`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[3]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testNodeList() {
    const testName = "testNodeList";
    const defaultOutput = "NodeList [ button#btn-1.btn.btn-log, " +
      "button#btn-2.btn.btn-err, button#btn-3.btn.btn-count ]";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[3]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testDocumentFragment() {
    const testName = "testDocumentFragment";

    const defaultOutput = "DocumentFragment [ li#li-0.list-element, " +
      "li#li-1.list-element, li#li-2.list-element, 2 more… ]";

    const longOutput = "DocumentFragment [ " +
      "li#li-0.list-element, li#li-1.list-element, li#li-2.list-element, " +
      "li#li-3.list-element, li#li-4.list-element ]";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `[5]`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: longOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testOnDomNodeMouseOver() {
    const stub = getGripStub("testNodeList");
    const grips = getSelectableInInspectorGrips(stub);

    is(grips.length, 3, "the stub has three node grips");

    let mouseOverValue;
    let onDOMNodeMouseOver = (object) => {
      mouseOverValue = object;
    };
    const renderedComponent = renderComponent(GripArray.rep, {
      object: stub,
      onDOMNodeMouseOver,
    });

    const nodes = renderedComponent.querySelectorAll(".objectBox-node");
    is(nodes.length, 3, "There are three node elements");
    nodes.forEach((node, index) => {
      TestUtils.Simulate.mouseOver(node);

      is(mouseOverValue, grips[index],
        "onDOMNodeMouseOver is called with the expected argument " +
        "when mouseover is fired on the Rep");
    });
  }

  function testOnDomNodeMouseOut() {
    const stub = getGripStub("testNodeList");
    const grips = getSelectableInInspectorGrips(stub);

    is(grips.length, 3, "the stub has three node grips");

    let called = 0;
    let onDOMNodeMouseOut = (object) => {
      called++;
    };
    const renderedComponent = renderComponent(GripArray.rep, {
      object: stub,
      onDOMNodeMouseOut,
    });

    const nodes = renderedComponent.querySelectorAll(".objectBox-node");
    info("Simulating mouseout on each node");
    Array.from(nodes).forEach(node => TestUtils.Simulate.mouseOut(node));

    is(called, 3, "onDOMNodeMouseOut is called when mouseout is fired on each NodeRep");
  }

  function testOnDomNodeInspectIconClick() {
    let inspectIconClickedValue = null;
    let onInspectIconClick = (object) => {
      inspectIconClickedValue = object;
    };

    let renderedComponentWithoutInspectIcon = renderComponent(GripArray.rep, {
      object: getGripStub("testDisconnectedNodeList"),
      onInspectIconClick,
    });
    is(renderedComponentWithoutInspectIcon.querySelector(".open-inspector"), null,
      "There isn't an inspect icon when the nodes are not connected to the DOM tree");

    const stub = getGripStub("testNodeList");
    const grips = getSelectableInInspectorGrips(stub);

    is(grips.length, 3, "the stub has three node grips");
    const renderedComponent = renderComponent(GripArray.rep, {
      object: stub,
      onInspectIconClick,
    });

    const icons = renderedComponent.querySelectorAll(".open-inspector");
    is(icons.length, grips.length,
      "There is an icon for each node connected to the DOM tree");

    icons.forEach((icon, index) => {
      TestUtils.Simulate.click(icon);

      is(inspectIconClickedValue, grips[index],
        "onInspectIconClick is called with the expected argument " +
        "when the inspect icon is clicked");
    });
  }

  function testObjectLink() {
    // Test array = `["test string"…] //4 items`
    const defaultOutput = `*Array **[ *${Array(maxLength.short).fill("\"test string\"").join(", ")}, *1 more…** ]*`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `*[*${maxLength.short + 1}*]*`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: `*Array **[ *${Array(maxLength.short + 1).fill("\"test string\"").join(", ")}* ]*`,
      }
    ];

    testRepRenderModes(
      modeTests,
      "testObjectLink",
      componentUnderTest,
      getGripStub("testMoreThanShortMaxProps"),
      {
        objectLink: (props, ...children) => React.DOM.span({},
          "*", ...children, "*"),
      }
    );
  }

  function getGripStub(functionName) {
    switch (functionName) {
      case "testBasic":
        return {
          "type": "object",
          "class": "Array",
          "actor": "server1.conn0.obj35",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 1,
          "preview": {
            "kind": "ArrayLike",
            "length": 0,
            "items": []
          }
        };

      case "testMaxProps":
        return {
          "type": "object",
          "class": "Array",
          "actor": "server1.conn1.obj35",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 4,
          "preview": {
            "kind": "ArrayLike",
            "length": 3,
            "items": [
              1,
              "foo",
              {
                "type": "object",
                "class": "Object",
                "actor": "server1.conn1.obj36",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0
              }
            ]
          }
        };

      case "testMoreThanShortMaxProps":
        let shortArrayGrip = {
          "type": "object",
          "class": "Array",
          "actor": "server1.conn1.obj35",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 4,
          "preview": {
            "kind": "ArrayLike",
            "length": maxLength.short + 1,
            "items": []
          }
        };

        // Generate array grip with length 4, which is more that the maximum
        // limit in case of the 'short' mode.
        for (let i = 0; i < maxLength.short + 1; i++) {
          shortArrayGrip.preview.items.push("test string");
        }

        return shortArrayGrip;

      case "testMoreThanLongMaxProps":
        let longArrayGrip = {
          "type": "object",
          "class": "Array",
          "actor": "server1.conn1.obj35",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 4,
          "preview": {
            "kind": "ArrayLike",
            "length": maxLength.long + 1,
            "items": []
          }
        };

        // Generate array grip with length 301, which is more that the maximum
        // limit in case of the 'long' mode.
        for (let i = 0; i < maxLength.long + 1; i++) {
          longArrayGrip.preview.items.push("test string");
        }

        return longArrayGrip;

      case "testPreviewLimit":
        return {
          "type": "object",
          "class": "Array",
          "actor": "server1.conn1.obj31",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 12,
          "preview": {
            "kind": "ArrayLike",
            "length": 11,
            "items": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
          }
        };

      case "testRecursiveArray":
        return {
          "type": "object",
          "class": "Array",
          "actor": "server1.conn3.obj42",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 2,
          "preview": {
            "kind": "ArrayLike",
            "length": 1,
            "items": [
              {
                "type": "object",
                "class": "Array",
                "actor": "server1.conn3.obj43",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 2,
                "preview": {
                  "kind": "ArrayLike",
                  "length": 1
                }
              }
            ]
          }
        };

      case "testNamedNodeMap":
        return {
          "type": "object",
          "class": "NamedNodeMap",
          "actor": "server1.conn3.obj42",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 6,
          "preview": {
            "kind": "ArrayLike",
            "length": 3,
            "items": [
              {
                "type": "object",
                "class": "Attr",
                "actor": "server1.conn3.obj43",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 2,
                  "nodeName": "class",
                  "value": "myclass"
                }
              },
              {
                "type": "object",
                "class": "Attr",
                "actor": "server1.conn3.obj44",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 2,
                  "nodeName": "cellpadding",
                  "value": "7"
                }
              },
              {
                "type": "object",
                "class": "Attr",
                "actor": "server1.conn3.obj44",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 2,
                  "nodeName": "border",
                  "value": "3"
                }
              }
            ]
          }
        };

      case "testNodeList":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj51",
          "class": "NodeList",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 3,
          "preview": {
            "kind": "ArrayLike",
            "length": 3,
            "items": [
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj52",
                "class": "HTMLButtonElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "button",
                  "isConnected": true,
                  "attributes": {
                    "id": "btn-1",
                    "class": "btn btn-log",
                    "type": "button"
                  },
                  "attributesLength": 3
                }
              },
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj53",
                "class": "HTMLButtonElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "button",
                  "isConnected": true,
                  "attributes": {
                    "id": "btn-2",
                    "class": "btn btn-err",
                    "type": "button"
                  },
                  "attributesLength": 3
                }
              },
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj54",
                "class": "HTMLButtonElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "button",
                  "isConnected": true,
                  "attributes": {
                    "id": "btn-3",
                    "class": "btn btn-count",
                    "type": "button"
                  },
                  "attributesLength": 3
                }
              }
            ]
          }
        };

      case "testDisconnectedNodeList":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj51",
          "class": "NodeList",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 3,
          "preview": {
            "kind": "ArrayLike",
            "length": 3,
            "items": [
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj52",
                "class": "HTMLButtonElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "button",
                  "isConnected": false,
                  "attributes": {
                    "id": "btn-1",
                    "class": "btn btn-log",
                    "type": "button"
                  },
                  "attributesLength": 3
                }
              },
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj53",
                "class": "HTMLButtonElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "button",
                  "isConnected": false,
                  "attributes": {
                    "id": "btn-2",
                    "class": "btn btn-err",
                    "type": "button"
                  },
                  "attributesLength": 3
                }
              },
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj54",
                "class": "HTMLButtonElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "button",
                  "isConnected": false,
                  "attributes": {
                    "id": "btn-3",
                    "class": "btn btn-count",
                    "type": "button"
                  },
                  "attributesLength": 3
                }
              }
            ]
          }
        };

      case "testDocumentFragment":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj45",
          "class": "DocumentFragment",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 0,
          "preview": {
            "kind": "DOMNode",
            "nodeType": 11,
            "nodeName": "#document-fragment",
            "childNodesLength": 5,
            "childNodes": [
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj46",
                "class": "HTMLLIElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "li",
                  "attributes": {
                    "id": "li-0",
                    "class": "list-element"
                  },
                  "attributesLength": 2
                }
              },
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj47",
                "class": "HTMLLIElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "li",
                  "attributes": {
                    "id": "li-1",
                    "class": "list-element"
                  },
                  "attributesLength": 2
                }
              },
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj48",
                "class": "HTMLLIElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "li",
                  "attributes": {
                    "id": "li-2",
                    "class": "list-element"
                  },
                  "attributesLength": 2
                }
              },
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj49",
                "class": "HTMLLIElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "li",
                  "attributes": {
                    "id": "li-3",
                    "class": "list-element"
                  },
                  "attributesLength": 2
                }
              },
              {
                "type": "object",
                "actor": "server1.conn1.child1/obj50",
                "class": "HTMLLIElement",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0,
                "preview": {
                  "kind": "DOMNode",
                  "nodeType": 1,
                  "nodeName": "li",
                  "attributes": {
                    "id": "li-4",
                    "class": "list-element"
                  },
                  "attributesLength": 2
                }
              }
            ]
          }
        };
      case "Array(5)" :
        return {
          "type": "object",
          "actor": "server1.conn4.child1/obj33",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 1,
          "preview": {
            "kind": "ArrayLike",
            "length": 5,
            "items": [
              null,
              null,
              null,
              null,
              null
            ]
          }
        };
      case "[,1,2,3]" :
        return {
          "type": "object",
          "actor": "server1.conn4.child1/obj35",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 4,
          "preview": {
            "kind": "ArrayLike",
            "length": 4,
            "items": [
              null,
              1,
              2,
              3
            ]
          }
        };
      case "[,,,3,4,5]" :
        return {
          "type": "object",
          "actor": "server1.conn4.child1/obj37",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 4,
          "preview": {
            "kind": "ArrayLike",
            "length": 6,
            "items": [
              null,
              null,
              null,
              3,
              4,
              5
            ]
          }
        };
      case "[0,1,,3,4,5]" :
        return {
          "type": "object",
          "actor": "server1.conn4.child1/obj65",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 6,
          "preview": {
            "kind": "ArrayLike",
            "length": 6,
            "items": [
              0,
              1,
              null,
              3,
              4,
              5
            ]
          }
        };
      case "[0,1,,,,5]" :
        return {
          "type": "object",
          "actor": "server1.conn4.child1/obj83",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 4,
          "preview": {
            "kind": "ArrayLike",
            "length": 6,
            "items": [
              0,
              1,
              null,
              null,
              null,
              5
            ]
          }
        };
      case "[0,,2,,4,5]" :
        return {
          "type": "object",
          "actor": "server1.conn4.child1/obj85",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 5,
          "preview": {
            "kind": "ArrayLike",
            "length": 6,
            "items": [
              0,
              null,
              2,
              null,
              4,
              5
            ]
          }
        };
      case "[0,,,3,,,,7,8]" :
        return {
          "type": "object",
          "actor": "server1.conn4.child1/obj87",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 5,
          "preview": {
            "kind": "ArrayLike",
            "length": 9,
            "items": [
              0,
              null,
              null,
              3,
              null,
              null,
              null,
              7,
              8
            ]
          }
        };
      case "[0,1,2,3,4,,]" :
        return {
          "type": "object",
          "actor": "server1.conn4.child1/obj89",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 6,
          "preview": {
            "kind": "ArrayLike",
            "length": 6,
            "items": [
              0,
              1,
              2,
              3,
              4,
              null
            ]
          }
        };
      case "[0,1,2,,,,]" :
        return {
          "type": "object",
          "actor": "server1.conn13.child1/obj88",
          "class": "Array",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 4,
          "preview": {
            "kind": "ArrayLike",
            "length": 6,
            "items": [
              0,
              1,
              2,
              null,
              null,
              null
            ]
          }
        };
    }
    return null;
  }
});
</script>
</pre>
</body>
</html>
